LayoutTests/fast/exclusions/parsing-wrap-shape-expected.txt

Testing the parsing of the -webkit-wrap-shape property.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS testCSSText("auto") is "auto"
PASS testComputedStyle("auto") is "auto"
PASS testCSSText("rect(10px, 20px, 30px, 40px)") is "rect(10px, 20px, 30px, 40px)"
PASS testComputedStyle("rect(10px, 20px, 30px, 40px)") is "rect(10px, 20px, 30px, 40px)"
PASS testCSSText("rect(10px, 20px, 30px, 40px, 5px)") is "rect(10px, 20px, 30px, 40px, 5px)"
PASS testComputedStyle("rect(10px, 20px, 30px, 40px, 5px)") is "rect(10px, 20px, 30px, 40px, 5px)"
PASS testCSSText("rect(10px, 20px, 30px, 40px, 5px, 10px)") is "rect(10px, 20px, 30px, 40px, 5px, 10px)"
PASS testComputedStyle("rect(10px, 20px, 30px, 40px, 5px, 10px)") is "rect(10px, 20px, 30px, 40px, 5px, 10px)"
PASS testCSSText("circle(10px, 20px, 30px)") is "circle(10px, 20px, 30px)"
PASS testComputedStyle("circle(10px, 20px, 30px)") is "circle(10px, 20px, 30px)"
PASS testCSSText("ellipse(10px, 20px, 30px, 40px)") is "ellipse(10px, 20px, 30px, 40px)"
PASS testComputedStyle("ellipse(10px, 20px, 30px, 40px)") is "ellipse(10px, 20px, 30px, 40px)"
PASS testCSSText("polygon(10px, 20px 30px, 40px 40px, 50px)") is "polygon(nonzero, 10px, 20px 30px, 40px 40px, 50px)"
PASS testComputedStyle("polygon(10px, 20px 30px, 40px 40px, 50px)") is "polygon(nonzero, 10px, 20px 30px, 40px 40px, 50px)"
PASS testCSSText("polygon(evenodd, 10px, 20px 30px, 40px 40px, 50px)") is "polygon(evenodd, 10px, 20px 30px, 40px 40px, 50px)"
PASS testComputedStyle("polygon(evenodd, 10px, 20px 30px, 40px 40px, 50px)") is "polygon(evenodd, 10px, 20px 30px, 40px 40px, 50px)"
PASS testCSSText("polygon(nonzero, 10px, 20px 30px, 40px 40px, 50px)") is "polygon(nonzero, 10px, 20px 30px, 40px 40px, 50px)"
PASS testComputedStyle("polygon(nonzero, 10px, 20px 30px, 40px 40px, 50px)") is "polygon(nonzero, 10px, 20px 30px, 40px 40px, 50px)"
PASS testNotInherited("auto", "rect(10px, 20px, 30px, 40px)") is "parent: auto, child: rect(10px, 20px, 30px, 40px)"
PASS testNotInherited("rect(10px, 20px, 30px, 40px)", "initial") is "parent: rect(10px, 20px, 30px, 40px), child: auto"
PASS testNotInherited("rect(10px, 20px, 30px, 40px)", "") is "parent: rect(10px, 20px, 30px, 40px), child: auto"
PASS testNotInherited("rect(10px, 20px, 30px, 40px)", "inherit") is "parent: rect(10px, 20px, 30px, 40px), child: rect(10px, 20px, 30px, 40px)"
PASS testNotInherited("", "inherit") is "parent: auto, child: auto"
PASS testNotInherited("auto", "inherit") is "parent: auto, child: auto"
PASS testCSSText("calc()") is ""
PASS testComputedStyle("calc()") is "auto"
PASS testCSSText("none") is ""
PASS testComputedStyle("none") is "auto"
PASS testCSSText("rect()") is ""
PASS testComputedStyle("rect()") is "auto"
PASS testCSSText("rect(10px)") is ""
PASS testComputedStyle("rect(10px)") is "auto"
PASS testCSSText("rect(10px, 10px)") is ""
PASS testComputedStyle("rect(10px, 10px)") is "auto"
PASS testCSSText("rect(10px, 20px, 30px)") is ""
PASS testComputedStyle("rect(10px, 20px, 30px)") is "auto"
PASS testCSSText("rect(10px 20px 30px 40px)") is ""
PASS testComputedStyle("rect(10px 20px 30px 40px)") is "auto"
PASS testCSSText("rect(10px, 20px, 30px, 40px, 50px, 60px, 70px)") is ""
PASS testComputedStyle("rect(10px, 20px, 30px, 40px, 50px, 60px, 70px)") is "auto"
PASS testCSSText("circle()") is ""
PASS testComputedStyle("circle()") is "auto"
PASS testCSSText("circle(10px)") is ""
PASS testComputedStyle("circle(10px)") is "auto"
PASS testCSSText("circle(10px, 20px)") is ""
PASS testComputedStyle("circle(10px, 20px)") is "auto"
PASS testCSSText("circle(10px 20px 30px)") is ""
PASS testComputedStyle("circle(10px 20px 30px)") is "auto"
PASS testCSSText("circle(10px, 20px, 30px, 40px)") is ""
PASS testComputedStyle("circle(10px, 20px, 30px, 40px)") is "auto"
PASS testCSSText("ellipse()") is ""
PASS testComputedStyle("ellipse()") is "auto"
PASS testCSSText("ellipse(10px)") is ""
PASS testComputedStyle("ellipse(10px)") is "auto"
PASS testCSSText("ellipse(10px, 20px)") is ""
PASS testComputedStyle("ellipse(10px, 20px)") is "auto"
PASS testCSSText("ellipse(10px, 20px, 30px)") is ""
PASS testComputedStyle("ellipse(10px, 20px, 30px)") is "auto"
PASS testCSSText("ellipse(10px 20px 30px 40px)") is ""
PASS testComputedStyle("ellipse(10px 20px 30px 40px)") is "auto"
PASS testCSSText("polygon()") is ""
PASS testComputedStyle("polygon()") is "auto"
PASS testCSSText("polygon(evenodd 10px, 20px 30px, 40px 40px, 50px)") is ""
PASS testComputedStyle("polygon(evenodd 10px, 20px 30px, 40px 40px, 50px)") is "auto"
PASS testCSSText("polygon(nonzero 10px, 20px 30px, 40px 40px, 50px)") is ""
PASS testComputedStyle("polygon(nonzero 10px, 20px 30px, 40px 40px, 50px)") is "auto"
PASS testCSSText("polygon(nonzero)") is ""
PASS testComputedStyle("polygon(nonzero)") is "auto"
PASS testCSSText("polygon(evenodd)") is ""
PASS testComputedStyle("polygon(evenodd)") is "auto"
PASS testCSSText("polygon(10px)") is ""
PASS testComputedStyle("polygon(10px)") is "auto"
PASS testCSSText("polygon(nonzero,10px)") is ""
PASS testComputedStyle("polygon(nonzero,10px)") is "auto"
PASS testCSSText("polygon(evenodd,12px)") is ""
PASS testComputedStyle("polygon(evenodd,12px)") is "auto"
PASS testCSSText("polygon(10px, 20px, 30px, 40px, 40px, 50px)") is ""
PASS testComputedStyle("polygon(10px, 20px, 30px, 40px, 40px, 50px)") is "auto"
PASS successfullyParsed is true

TEST COMPLETE
